Update OpenApi, fix models renaming
